refactor(webapp): preparing for physiotherapist profile page
