+++ webapp/src/components/physiotherapist/PhysiotherapistInfo.tsx
@@ -1,0 +1,7 @@
+export interface PhysiotherapistInfoProps {
+  physiotherapist: string;
+}
+
+export default function PhysiotherapistInfo(props: PhysiotherapistInfoProps) {
+  return <div>{props.physiotherapist}</div>;
+}
